fix dtype

+++ TEST/TagReadWithGeneExon/test/scripts/test.py
@@ -32,7 +32,7 @@ for ref in reads_dict:
     for read in reads_list:
         total += len(read.get_blocks())
 
-    R = np.zeros(total)
+    R = np.zeros(total, dtype='i16')
     B = np.arange(total)
 
     ctr = 0
